pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opsbeacon"
version = "1.3.0"
description = "OpsBeacon Python SDK - Client library for interacting with the OpsBeacon API"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Cihan Sahin", email = "cihan@opsbeacon.com" }
]
maintainers = [
    { name = "OB2 AI", email = "support@opsbeacon.com" }
]
keywords = [
    "opsbeacon",
    "api",
    "sdk",
    "automation",
    "devops",
    "mcp",
    "model-context-protocol"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.25.1",
]

[project.optional-dependencies]
cli = [
    "python-dotenv>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "responses>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-requests>=2.28.0",
    "pre-commit>=3.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
]
all = [
    "opsbeacon[cli,dev,docs]",
]

[project.urls]
Homepage = "https://opsbeacon.com"
Documentation = "https://github.com/ob2ai/ob-python-sdk#readme"
Repository = "https://github.com/ob2ai/ob-python-sdk"
Issues = "https://github.com/ob2ai/ob-python-sdk/issues"
Changelog = "https://github.com/ob2ai/ob-python-sdk/blob/main/CHANGELOG.md"

[project.scripts]
opsbeacon = "opsbeacon.cli:main"

# Hatch build configuration
[tool.hatch.build.targets.sdist]
include = [
    "/opsbeacon",
    "/tests",
    "/README.md",
    "/LICENSE",
    "/CHANGELOG.md",
]

[tool.hatch.build.targets.wheel]
packages = ["opsbeacon"]

# Ruff configuration (linting + formatting)
[tool.ruff]
target-version = "py38"
line-length = 100
exclude = [
    ".git",
    ".ruff_cache",
    ".mypy_cache",
    ".pytest_cache",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # pylint
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",    # line too long (handled by formatter)
    "PLR0913", # too many arguments
    "PLR2004", # magic value comparison
    "UP007",   # Use X | Y instead of Union - need Union for Python 3.8 compatibility
    "UP045",   # Use X | None instead of Optional - need Optional for Python 3.8 compatibility
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "PLR2004", "PLC0415"]
"examples/*" = ["ARG", "ERA"]

[tool.ruff.lint.isort]
known-first-party = ["opsbeacon"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# MyPy configuration
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=opsbeacon",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=40",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests (require API access)",
    "slow: Slow running tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["opsbeacon"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
